apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg2-binary>=2.9.10
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
